refactor(ethportal-api): reth type funcs to module
The type conversion functions will be reused in several places, and
don't belong co-located with the header types.
