+++ ethportal-api/src/types/execution/header.rs
@@ -1,13 +1,13 @@
 use ethereum_types::{Bloom, H160, H256, H64, U256};
 use reth_rpc_types::Header as RpcHeader;
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
-use ruint::Uint;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use ssz::{Encode, SszDecoderBuilder, SszEncoder};
 use ssz_derive::{Decode, Encode};
 
 use crate::types::bytes::ByteList;
 use crate::utils::bytes::{hex_decode, hex_encode};
+use crate::utils::rethtypes::{u256_to_uint256, u64_to_uint256};
 
 const LONDON_BLOCK_NUMBER: u64 = 12965000;
 const SHANGHAI_BLOCK_NUMBER: u64 = 17034871;
@@ -256,22 +256,6 @@ impl From<Header> for RpcHeader {
             parent_beacon_block_root: None,
         }
     }
-}
-
-fn u256_to_uint256(u256: U256) -> Uint<256, 4> {
-    let mut bytes = [0u8; 32];
-    u256.to_big_endian(&mut bytes);
-    Uint::from_be_bytes(bytes)
-}
-
-fn u64_to_uint256(val: u64) -> Uint<256, 4> {
-    let u64_bytes: &[u8] = &val.to_be_bytes();
-    let high_zero_bytes: &[u8] = &[0u8; 24];
-    let bytes: [u8; 32] = [high_zero_bytes, u64_bytes]
-        .concat()
-        .try_into()
-        .expect("8 bytes + 24 bytes should be 32 bytes");
-    Uint::from_be_bytes(bytes)
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
