+++ ethportal-api/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod bytes;
+pub mod rethtypes;
 pub mod serde;
